pyproject.toml
[project]
name = "devx-cli"
version = "0.2.0"
description = "DevX CLI modular: health, loadtest, linkscan, secrets, docgen"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.3",
  "markdownify>=0.13.1",
  "python-dotenv>=1.0.1"
]

[project.scripts]
devx = "devx.cli:main"
